pyproject.toml
[project]
name = "basics-of-blockchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "py-solc-x>=2.0.5",
    "web3>=7.14.1",
]
